Add a separate controller for accreditation
