+++ src/types/swagger-api-types.ts
@@ -332,6 +332,10 @@
  *       type: object
  *       additionalProperties: false
  *       properties:
+ *         issuerDid:
+ *           description: DID of the Verifiable Credential issuer. This needs to be a `did:cheqd` DID.
+ *           type: string
+ *           example: did:cheqd:testnet:7bf81a20-633c-4cc7-bc4a-5a45801005e0
  *         subjectDid:
  *           description: DID of the Verifiable Credential holder/subject. This needs to be a `did:key` DID.
  *           type: string
@@ -441,9 +445,11 @@
  *           - verida
  *           - resource
  *       required:
+ *         - issuerDid
  *         - subjectDid
  *         - schemas
  *       example:
+ *         issuerDid: did:cheqd:testnet:7bf81a20-633c-4cc7-bc4a-5a45801005e0
  *         subjectDid: did:key:z6MkhaXgBZDvotDkL5257faiztiGiC2QtKLGpbnnEGta2doK
  *         schemas:
  *           - type: MuseumPassCredential
